requirements.txt
django==5.0.8
django-celery-beat

psycopg2-binary
django-redis

python-dotenv~=1.0.1
django-modeltranslation~=0.19.9